restrict save on invalid url n update test cases
